# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Agency Owner Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDetail"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Projects ────────────────────────────────────────────────────────────
...
  /clients/{clientId}/documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents for a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create a document/link for a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    patch:
      operationId: updateDocument
      tags: [documents]
      summary: Update a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentUpdate"
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Notes ───────────────────────────────────────────────────────────────
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a single notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "404":
          description: Not found
  /notifications/{id}:
    delete:
      operationId: deleteNotification
      tags: [notifications]
      summary: Delete a notification
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ─── Admin ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Client:
      type: object
      required: [id, companyName, status, createdAt]
      properties:
        id:
          type: integer
        companyName:
          type: string
        logoUrl:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        primaryContact:
          type: ["string", "null"]
        secondaryContact:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        timezone:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, prospect, churned]
        startDate:
          type: ["string", "null"]
        contractValue:
          type: ["number", "null"]
        monthlyRetainer:
          type: ["number", "null"]
        paymentMethod:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    ClientDetail:
      allOf:
        - $ref: "#/components/schemas/Client"
        - type: object
          properties:
            projects:
              type: array
              items:
                $ref: "#/components/schemas/Project"
            recentActivity:
              type: array
              items:
                $ref: "#/components/schemas/ActivityItem"
            openPayments:
              type: array
              items:
                $ref: "#/components/schemas/Payment"
            totalRevenue:
              type: number
            outstandingBalance:
              type: number
    ClientInput:
      type: object
      required: [companyName]
      properties:
        companyName:
          type: string
        logoUrl:
          type: string
        industry:
          type: string
        website:
          type: string
        email:
          type: string
        phone:
          type: string
        primaryContact:
          type: string
        secondaryContact:
          type: string
        address:
          type: string
        timezone:
          type: string
        status:
          type: string
          enum: [active, inactive, prospect, churned]
        startDate:
          type: string
        contractValue:
          type: number
        monthlyRetainer:
          type: number
        paymentMethod:
          type: string
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
    ClientUpdate:
      type: object
      properties:
        companyName:
          type: string
        logoUrl:
          type: string
        industry:
          type: string
        website:
          type: string
        email:
          type: string
        phone:
          type: string
        primaryContact:
          type: string
        secondaryContact:
          type: string
        address:
          type: string
        timezone:
          type: string
        status:
          type: string
          enum: [active, inactive, prospect, churned]
        startDate:
          type: string
        contractValue:
          type: number
        monthlyRetainer:
          type: number
        paymentMethod:
          type: string
        notes:
          type: string
        tags:
          type: array
          items:
            type: string

    # ─── Project ─────────────────────────────────────────────────────────────
    Project:
      type: object
      required: [id, clientId, name, status, priority, progress, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        name:
          type: string
        status:
          type: string
          enum: [planning, waiting, design, development, testing, review, delivered, paused, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        startDate:
          type: ["string", "null"]
        deadline:
          type: ["string", "null"]
        estimatedBudget:
          type: ["number", "null"]
        actualCost:
          type: ["number", "null"]
        revenue:
          type: ["number", "null"]
        profit:
          type: ["number", "null"]
        description:
          type: ["string", "null"]
        ownerNotes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    Payment:
      type: object
      required: [id, clientId, invoiceNumber, amount, status, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        projectId:
          type: ["integer", "null"]
        invoiceNumber:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [paid, pending, overdue, cancelled]
        dueDate:
          type: ["string", "null"]
        paidDate:
          type: ["string", "null"]
        paymentMethod:
          type: ["string", "null"]
        remainingBalance:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Document:
      type: object
      required: [id, clientId, title, type, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        projectId:
          type: ["integer", "null"]
        title:
          type: string
        type:
          type: string
          enum: [contract, invoice, proposal, design, brand_assets, link, google_drive, github, figma, other]
        url:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    DocumentInput:
      type: object
      required: [title, type]
      properties:
        projectId:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [contract, invoice, proposal, design, brand_assets, link, google_drive, github, figma, other]
        url:
          type: string
        notes:
          type: string
    DocumentUpdate:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [contract, invoice, proposal, design, brand_assets, link, google_drive, github, figma, other]
        url:
          type: string
        notes:
          type: string

    # ─── Note ────────────────────────────────────────────────────────────────
...
    ActivityItem:
      type: object
      required: [id, type, description, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        entityType:
          type: string
        entityId:
          type: ["integer", "null"]
        description:
          type: string
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        createdAt:
          type: string

    # ─── Calendar ────────────────────────────────────────────────────────────
...
    Notification:
      type: object
      required: [id, type, title, message, entityType, isRead, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        title:
          type: string
        message:
          type: string
        entityType:
          type: string
        entityId:
          type: ["integer", "null"]
        href:
          type: ["string", "null"]
        isRead:
          type: boolean
        createdAt:
          type: string
...