# lib/api-zod/src/generated/types/notification.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Agency Owner Operating System API
 * OpenAPI spec version: 0.1.0
 */

export interface Notification {
  id: number;
  type: string;
  title: string;
  message: string;
  entityType: string;
  /** @nullable */
  entityId?: number | null;
  /** @nullable */
  href?: string | null;
  isRead: boolean;
  createdAt: string;
}